+++ Content.Shared/_Moffstation/Medical/AdvancedCryogenics/CryocapsuleComponent.cs
@@ -1,4 +1,5 @@
 using Content.Shared.Containers.ItemSlots;
+using Robust.Shared.Serialization;
 
 namespace Content.Shared._Moffstation.Medical.AdvancedCryogenics;
 
@@ -8,15 +9,61 @@ namespace Content.Shared._Moffstation.Medical.AdvancedCryogenics;
 [RegisterComponent]
 public sealed partial class CryocapsuleComponent : Component
 {
-    /// <summary>
-    /// The ID of the itemslot that holds the cryo unit
-    /// </summary>
-    [DataField]
-    public string CasketSlotId = "brain_slot";
+    #region organs slots
 
     /// <summary>
-    /// The <see cref="ItemSlot"/> for this capsule. Holds the brain.
+    /// The ID of the itemslot that holds the brain.
+    /// </summary>
+    [DataField]
+    public string BrainSlotId = "brain_slot";
+
+    /// <summary>
+    /// The <see cref="ItemSlot"/> that holds the brain.
     /// </summary>
     [DataField(required: true)]
-    public ItemSlot CasketSlot = new();
+    public ItemSlot BrainSlot = new();
+
+    /// <summary>
+    /// The ID of the itemslot that holds the lungs.
+    /// </summary>
+    [DataField]
+    public string LungSlotId = "lung_slot";
+
+    /// <summary>
+    /// The <see cref="ItemSlot"/> that holds the lungs.
+    /// </summary>
+    [DataField]
+    public ItemSlot LungSlot = new();
+
+
+    /// <summary>
+    /// The ID of the itemslot that holds the stomach.
+    /// </summary>
+    [DataField]
+    public string StomachSlotId = "stomach_slot";
+
+    /// <summary>
+    /// The <see cref="ItemSlot"/> that holds the stomach.
+    /// </summary>
+    [DataField]
+    public ItemSlot StomachSlot = new();
+
+    #endregion
+}
+
+[Serializable, NetSerializable]
+public enum CryocasketVisuals : byte
+{
+    BrainPresent,
+    HasMind
+}
+
+[Serializable, NetSerializable]
+public enum CryocasketVisualLayers : byte
+{
+    Brain,
+    Lungs,
+    Heart,
+    Stomach,
+    Base
 }
