--- Content.Shared/_Moffstation/Medical/AdvancedCryogenics/CryocasketSystem.cs
@@ -1,26 +1,0 @@
-using Content.Shared.Containers.ItemSlots;
-
-namespace Content.Shared._Moffstation.Medical.AdvancedCryogenics;
-
-/// <summary>
-/// This handles...
-/// </summary>
-public sealed class CryocasketSystem : EntitySystem
-{
-    [Dependency] private readonly ItemSlotsSystem _itemSlots = default!;
-
-    /// <inheritdoc/>
-    public override void Initialize()
-    {
-        base.Initialize();
-
-        SubscribeLocalEvent<CryoCasketComponent, ComponentInit>(OnCryocasketInit);
-    }
-
-    private void OnCryocasketInit(Entity<CryoCasketComponent> ent, ref ComponentInit init)
-    {
-        _itemSlots.AddItemSlot(ent.Owner, ent.Comp.BrainSlotId, ent.Comp.BrainSlot);
-        _itemSlots.AddItemSlot(ent.Owner, ent.Comp.LungSlotId, ent.Comp.LungSlot);
-        _itemSlots.AddItemSlot(ent.Owner, ent.Comp.StomachSlotId, ent.Comp.StomachSlot);
-    }
-}
