+++ Resources/Locale/en-US/_Moffstation/medical/components/cryocapsule-component.ftl
@@ -1,0 +1,7 @@
+# Slots names
+cryocapsule-brain-slot-name = Brain
+cryocapsule-lungs-slot-name = Lungs
+cryocapsule-stomach-slot-name = Stomach
+
+cryomachine-capsule-slot-name = Cryocapsule
+cryomachine-beaker-slot-name = Solution
